+++ examples/demo-rollup/tests/restart/mod.rs
@@ -89,6 +89,7 @@ async fn start_stop_empty(
             )
             .with_zkvm_host_args(mock_da_risc0_host_args())
             .set_config(|c| {
+                c.max_concurrent_blobs = 65536;
                 c.storage = rollup_storage_dir.clone();
                 c.rollup_prover_config = Some(rollup_prover_config.clone());
                 c.aggregated_proof_block_jump = 10;
@@ -215,6 +216,7 @@ async fn test_start_prover_manual() -> anyhow::Result<()> {
     )
     .with_zkvm_host_args(mock_da_risc0_host_args())
     .set_config(|c| {
+        c.max_concurrent_blobs = 65536;
         c.storage = rollup_storage_dir.clone();
         c.rollup_prover_config = Some(RollupProverConfig::Skip);
         c.aggregated_proof_block_jump = jump_size;
@@ -366,6 +368,7 @@ async fn check_with_increasing_stf_infos(
     )
     .with_zkvm_host_args(mock_da_risc0_host_args())
     .set_config(|c| {
+        c.max_concurrent_blobs = 65536;
         c.storage = rollup_storage_dir.clone();
         c.rollup_prover_config = Some(RollupProverConfig::Skip);
         c.aggregated_proof_block_jump = aggregated_proof_jump;